Reduce tcp port usage on master.
QD will establish TCP connections to each QE. QEs on the same segment have the same port and
address, thus, we need QE of different gang to have different QD port so that the connection
is a unique <source address, source port, destination address, destination port> quadruple.

To reduce the usage of TCP port on master. We could let the QEs of the same gang on different
segment share on same QD port.