Cleanup cdbdisp_handleError function where exists confusing codes.
Back to the old tcp interconnect time, connections are established between slices.
If the downstream slices report a no-interconnect error, the upstream will find the
connection is disconnect and may report a interconnect error such as "sender closed
connection prematurely" which is not the accurate error, so cdbdisp_handleError need
to promote error to avoid this. Since tcp interconnect is no longer existed in current
code base, and new udpifc has no such issue, it's time to abandon cdbdisp_handleError().