Clean up some cases from bugbuster
1. Remove TestPSQL from bugbuster
The whole test is completely enclosed in a start/end_ignore block, we can not
verify the correctness except it crash the whole service. Moreover,
allow_system_table_mods has been tested by other cases in ICG, so this test case
is useless for now.

2. Remove useless codecoverage case from bugbuster

3. Remove gp_persistent_table from bugbuster
gp_enable_inheritance is not invalid under current code base, so Remove this
case

4. merge DMLstmnt from bugbuster to icg