resgroup: bypass SET/SHOW command for resource group
SET or SHOW GUC won't take significant resource including cpu and
memory, we can safely let it run. It's useful that not be blocked
when changing from a busy group immediately to another group. For
later integration with WLM, it's also helpful.

We only check the type of current statement when there are no free
slots in the current group, so the overhead of parsing will not
affect normal procedure.