Ignore undetermined output within cursor.sql
When declaring a cursor, QD will not block until QE got a
snapshot and set up a interconnect, so we can not guarantee
that QD always see 'qe_got_snapshot_and_interconnect' was
triggered. For the case itself, if the writer QE can get
slower than the reader QE without the help of fault injector,
it's even better.