Remove incorrect dispatcher test
createGang() connected each primary segments to generate a gang,
if an error occurs, it triggers the FTS mechanism to test if the
segments is alive, we previously injected a faultinjector within
ProcessStartupPackets() to trigger an timeout error. The problem
now is that this faultinjector will also suspend the FTS prober
and the whole test get stuck.

FTS prober used to use the PRIMARY_MIRROR_TRANSITION_REQUEST_CODE
(a special protocol) way which didn't trigger faultinjector in
ProcessStartupPackets(), now it uses the walrep way and unexpected
triggered the faultinjector.