Fix flaky dispatcher test case
We added a fault injector to sleep 5 seconds in segment to simulate
a non-respond case so we can test GUC gp_segment_connect_timeout can
take effect when creating a gang. Previously, a reader gang triggered
the fault injector and it may report a FATAL error unexpectedly because
the writer gang was destroyed.

To fix this, we let a writer gang to trigger the fault injector and
let the case wait another 5 seconds so the writer gang can exit
normally.