Remove unused function 'isPrimaryWriterGangAlive'
It's originally added to test if the writer gang is alive
(like segment be killed by 9) by calling an actual recv()
syscall. It's removed because of calling syscall for each
segments is heavy, there is no need to do the optimization
for such a rare case, the error can bump up when the command
is actually sent.