Remove gang_id sliceid map
GPDB used to dispatch a sliceid+gang_id map to QE to tell which slice
a QE should execute, A gang_id is assigned when the QE is started, so
if a QE is reused by another gang, it can not get the right slice to
execute.

To do this, this commit put sliceid info in the queryText and dispath
it to QE, so we need to copy a original queryText and replace the
slice id accordingly.