Remove gang_id sliceid map
The commit revert 4b3609423ae8e1d269bf0b02238fd05af7ca495d which let
QD dispatch a sliceid + gang_id map to QE to tell which slice a QE
should execute, A gang_id is assigned when the QE is started, so
if a QE is reused by another gang, it can not get the right slice to
execute.

This is a temporary fix, once each QE in a session has a unique
identifier, we can also add a map for QE and slice, then there is
no need to copy query text for each slice.