Fix memory context bug
Previously, when waiting dispatcher to finished, it will first check
if there are outbound data in the connections, if outbound data exist,
dispatcher will resend it. A problem is those outbound data  might be
allocated in ExecutorContext which might be destroyed before dispatcher
state. To fix this, we need to allocated query text in the context of
dispatcher.