Add a slice to QE map so only one query copy is needed by dispatcher
We used to use slice to gang_id to avoid dispatching different copy
of query for each slice, however, QE now is not bound to a fixed gang,
so a new map is needed to replace it. The idea is CdbComponentDatabases
assign a unique ID to each QE worker and each slice maintain a bitmap
of worker ID, when a QE worker get a plan, it goes through the slice
table to find which slice it belongs to.