Fix mirror_promote test cases
This case manually stop the primary using pg_ctl which causes
the primaryWriterGang cache of session invalid, so when this
session trying to inject fts_handle_message fault, it will
find the connection with segment 0 is invalid.

Old code didn't expose this error because it called
freeGangsForPortal for every commands no matter the command
used primaryWriterGang or not.

This actually exposes the problem that current segments
management make segment status inconsistent with FTS status,
another PR will resolve this problem.