Do a primary writer gang validation at CommitTransaction
Previously, QD did a writer gang validation at the end of
CommitTransaction no matter this query used the gang or not.

A few test cases rely on this feature, to make test cases
happy, we add a validation back. This is only temp fix,
ideally, dispatcher can recreate a writer gang if available
writer gang is bad already in certain safty case.