Fix the interval issue when Seting up TCP interconnect
When TCP connections cannot be setup for a long time, we check
if some segments are already failed out, the check is an high-cost
operation, so we set the interval to 2 seconds. We used to use a
counter to record the interval which is not reliable because
a loop cycle (500ms) may be interrupted earlier due to
EINT/EAGAIN of select().

To not affect the setup performance of TCP interconnect, we need
to make the interval mechanism more reliable.