Fix incorrect assertion within adjust_modifytable_flow
Previously, when we try to update a partition table or an inherited table,
we assume all parent table and child table should have the same number of
segments in gp_distribution_policy catalog, otherwise, we let the query
fail.

The assumption is incorrect, updating on mix size of target relations
can work actually, we only need to set the flow size of modifytable node
to the maxinum size of all target relations and take care of the output
size of motion for each target relations individually.

It's very common that parent table and child table has different number
of segments in gp_distribution_policy, gpexpand may expand the child
table first and then the parent table.