Online expand improvement
commit 8eed421738 & e0b06678aa allow us to do a gpexpand of GPDB cluster
without a restart, so we called this strategy as "online expand", those
two commits were mainly focus on how to avoid restarting cluster when
expanding cluster, this commit will do the left work to improve gpexpand
using a few features we merged to master recently.

First improvement is, it's no longer necessary to change policy of all
non-partition table to random when doing expanding. Previously, we couldn't
tell the difference between expanded table and non-expaned table, so if
the distribution policy of these tables were the same, planner took these
tables's data as co-located and produced a incorrect plan. To avoid this,
gpexpand used to change policy of all table to random so planner could
produce a correct but a non-effective plan, because random policy always
means annoying broadcase motion.

Now each table has a 'numsegments' attribute introduced by 4eb65a53, GPDB
can recognize expanded and non-expand table and produced correct plans. so
the first imporvement is removing the randomization step of tables.

The second improvement is, use a brand new syntax "ALTER TABLE foo EXPAND TABLE"
to focus on the data rebalance of tables. Previously, tables were converted
to randomly before rebalance and numsegments of tables were always the same
with GPDB cluster size, so we can use a tricky "ALTER TABLE foo SET WITH
(REORGNIZE = true) DISTRIBUTED BY (original_key)" command to rebalance data
to new added segments, now policy and numsegments are not changed before
rebalancing data, so expanding such tables is not match the concept of "SET
DISTRIBUTED BY" now, we need a proper syntax to focus on table expanding.

New expand syntax named "ALTER TABLE foo EXPAND TABLE", we have two methods
to do the real data movement inside, one is CTAS, another is RESHUFFLE, which
one is better depends on how much data needs to be moved and whether the table
has index and whether the table is a append only table (analyzed by Hekki).

The third improvement is, postpone the randomization of partition tables.
Partition table stores data on leaf partitions only, previously, to expanded
leaf partitions in parallel, gpexpand randomize policy of leaf part only and
keep the policy of root/interior part no changed, SELECT/UPDATE/INSERT
operations on the partition table work well in this mode, even some leaf
partition was expaned and other leaf partitons were not.

We want to keep the mode to expand partition table in parallel but postpone
the randomization so planner can still use the policy info to produce
effective plan before partition table is actually expanded. Another thing
is we are online expand now, using utility mode directly is not safe, we need
to add lock protection and keep the consistency of gp_distribution_policy
when doing the randomization. As a result, "ALTER TABLE root_p EXPAND PREPARE"
is designed to do the randomization and optimizations are added to let
EXPAND PREPARE only be called one time for a partition table.